+++ .eslintignore
@@ -1,7 +1,5 @@
 node_modules
-package-lock.json
 pnpm-lock.yaml
-yarn.lock
 
 # build dirs
 build
