--- .secret-sample
@@ -1,1 +1,0 @@
-somesecret
