--- docker-compose.yaml
@@ -1,16 +1,0 @@
-version: '3.9'
-
-services:
-  template:
-    build: .
-    environment:
-      SECRET_FILE: /run/secrets/secret
-      FOO: ${FOO}
-    secrets:
-      - secret
-    env_file:
-      - .env-sample
-
-secrets:
-  secret:
-    file: .secret-sample
