+++ package-scripts.yaml
@@ -41,17 +41,6 @@ scripts:
     #    script: tsc && nps barrels && lint-staged
     hiddenFromHelp: true
 
-  start:
-    default:
-      script: nps build start.prod
-      description: Start in production mode
-    dev:
-      script: tsnd --cls --exit-child --ignore-watch node_modules --quiet  --respawn --rs --transpile-only -r tsconfig-paths/register src
-      description: Start in dev mode
-    prod:
-      script: NODE_ENV=production node -r module-alias/register dist
-      hiddenFromHelp: true
-
   test:
     default:
       script: nps 'test.once --watch --coverage=false'
