+++ package.json
@@ -8,9 +8,6 @@
     "prepare": "husky install",
     "setup": "pnpm 2> /dev/null || npm i -g pnpm && pnpm i"
   },
-  "_moduleAliases": {
-    "@mod": "dist/mod"
-  },
   "author": "r1oga",
   "license": "ISC",
   "devDependencies": {
@@ -20,6 +17,8 @@
     "@types/jest": "^29.2.3",
     "@types/js-yaml": "^4.0.5",
     "@types/supertest": "^2.0.12",
+    "barrelsby": "^2.5.1",
+    "concurrently": "^7.6.0",
     "husky": "^8.0.2",
     "is-ci": "^3.0.1",
     "jest": "^29.3.1",
@@ -33,6 +32,7 @@
     "jest-watch-typeahead": "^2.2.1",
     "lint-staged": "^13.1.0",
     "msw": "^0.49.2",
+    "nps": "^5.10.0",
     "reflect-metadata": "^0.1.13",
     "supertest": "^6.3.1",
     "ts-jest": "^29.0.3",
@@ -42,13 +42,7 @@
     "typescript": "^4.9.4"
   },
   "dependencies": {
-    "barrelsby": "^2.5.1",
-    "concurrently": "^7.6.0",
-    "cookie": "^0.5.0",
-    "js-yaml": "^4.1.0",
-    "module-alias": "^2.2.2",
-    "nps": "^5.10.0",
-    "wretch": "^2.3.0"
+    "js-yaml": "^4.1.0"
   },
   "engines": {
     "node": ">=19"
