+++ src/decorators/index.ts
@@ -1,5 +1,17 @@
-/**
- * @file Automatically generated by barrelsby.
- */
+import { Dune } from '../Dune'
 
-export * from './maybe-get-csrf-token'
+export const maybeGetCsrfToken = (
+  target: object,
+  propertyKey: string,
+  descriptor: PropertyDescriptor,
+) => {
+  const originalMethod = descriptor.value
+
+  descriptor.value = async function () {
+    if ((this as Dune).cookies.csrf === undefined)
+      await (this as Dune)['getCsrfToken']()
+    return originalMethod.apply(this)
+  }
+
+  return descriptor
+}
