+++ test/jest.unit.ts
@@ -1,7 +1,5 @@
 import type { JestConfigWithTsJest } from 'ts-jest'
 
-import { pathsToModuleNameMapper } from 'ts-jest'
-import { compilerOptions } from '../tsconfig.json'
 import common from './jest.common'
 
 const jestUnitConfig: JestConfigWithTsJest = {
@@ -19,9 +17,6 @@ const jestUnitConfig: JestConfigWithTsJest = {
   displayName: 'unit',
   moduleDirectories: ['node_modules', __dirname],
   moduleFileExtensions: ['ts', 'js', 'json'],
-  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, {
-    prefix: '<rootDir>/',
-  }),
   preset: 'ts-jest',
   setupFilesAfterEnv: [
     'jest-chain',
