+++ tsconfig.json
@@ -24,10 +24,6 @@
     "noUnusedLocals": false,
     "skipDefaultLibCheck": true,
     "skipLibCheck": true,
-    "importHelpers": true,
-    "paths": {
-      "@mod": ["src/mod"],
-      "@mod/*": ["src/mod/*"]
-    }
+    "importHelpers": true
   }
 }
